[Core] Unify profiler into one single file
