+++ kratos/utilities/profiler.h
@@ -14,6 +14,7 @@
 
 // Project includes
 #include "includes/code_location.h"
+#include "includes/kratos_parameters.h"
 
 // System includes
 #include <unordered_map>
@@ -23,13 +24,45 @@
 #include <optional>
 #include <list>
 #include <mutex>
-
+#include <cstdio>
+#include <cstring>
+#include <algorithm>
+#include <vector>
+#include <fstream>
+#include <sstream>
+#include <atomic>
+#include <limits>
 
 namespace Kratos::Internals {
 
+namespace {
+template <class TTimeUnit>
+std::string GetTimeUnit()
+{
+    KRATOS_ERROR << "Unsupported time unit";
+}
+
+template <>
+std::string GetTimeUnit<std::chrono::milliseconds>()
+{
+    return "ms";
+}
+
+template <>
+std::string GetTimeUnit<std::chrono::microseconds>()
+{
+    return "us";
+}
+
+template <>
+std::string GetTimeUnit<std::chrono::nanoseconds>()
+{
+    return "ns";
+}
+} // unnamed namespace
 
 template <class TTimeUnit>
-class KRATOS_API(KRATOS_CORE) Profiler
+class Profiler
 {
 private:
     /// @brief Absolute time type.
@@ -45,17 +78,39 @@ private:
     class Item
     {
     public:
-        explicit Item(CodeLocation&& rLocation);
+        explicit Item(CodeLocation&& rLocation)
+            : Item(0,                                                               // <== .mCallCount
+                  Duration(0),                                                     // <== .mCumulative
+                  Duration(std::numeric_limits<typename Duration::rep>::max()),    // <== .mMin
+                  Duration(0),                                                     // <== .mMax
+                  std::move(rLocation))                                            // <== .mLocation
+        {
+        }
 
     private:
         Item(std::size_t CallCount,
              Duration CumulativeDuration,
              Duration MinDuration,
              Duration MaxDuration,
-             CodeLocation&& rLocation);
+             CodeLocation&& rLocation)
+            : mRecursionLevel(0)
+            , mCallCount(CallCount)
+            , mCumulative(CumulativeDuration)
+            , mMin(MinDuration)
+            , mMax(MaxDuration)
+            , mLocation(std::move(rLocation))
+        {
+        }
 
         /// @brief Aggregate profiled data from another @ref Item in the same scope.
-        Item& operator+=(const Item& rOther);
+        Item& operator+=(const Item& rOther)
+        {
+            mCallCount += rOther.mCallCount;
+            mCumulative += rOther.mCumulative;
+            mMin = std::min(mMin, rOther.mMin);
+            mMax = std::max(mMax, rOther.mMax);
+            return *this;
+        }
 
     private:
         friend class Profiler;
@@ -93,7 +148,7 @@ private:
     struct SourceLocationEquality
     {
         bool operator()(const CodeLocation& rLhs,
-                        const CodeLocation& rRhs) const
+                       const CodeLocation& rRhs) const
         {
             return (rLhs.GetFileName() == rRhs.GetFileName())
                    && (rLhs.GetFunctionName() == rRhs.GetFunctionName());
@@ -105,12 +160,33 @@ public:
     class Scope
     {
     public:
-        ~Scope();
+        ~Scope()
+        {
+            const auto end = Clock::now();
+            const auto duration = std::chrono::duration_cast<Duration>(end - mBegin);
+            if (0 == mrItem.mRecursionLevel) {
+                ++mrItem.mCallCount;
+                mrItem.mCumulative += duration;
+                mrItem.mMin = std::min(mrItem.mMin, duration);
+                mrItem.mMax = std::max(mrItem.mMax, duration);
+            }
+            --mrItem.mRecursionLevel;
+        }
 
     private:
-        Scope(Item& rItem);
+        Scope(Item& rItem)
+            : mrItem(rItem)
+            , mBegin(Clock::now())
+        {
+            ++mrItem.mRecursionLevel;
+        }
 
-        Scope(Item& rItem, std::chrono::high_resolution_clock::time_point Begin);
+        Scope(Item& rItem, std::chrono::high_resolution_clock::time_point Begin)
+            : mrItem(rItem)
+            , mBegin(Begin)
+        {
+            ++mrItem.mRecursionLevel;
+        }
 
         Scope(Scope&&) = delete;
 
@@ -139,24 +215,147 @@ public:
     >;
 
 public:
-    Profiler();
+    Profiler()
+        : Profiler("kratos_profiler_output_" + GetTimeUnit<T>() + ".json")
+    {
+    }
+
+    Profiler(std::filesystem::path&& r_outputPath)
+        : mItemContainerMap()
+        , mItem(KRATOS_CODE_LOCATION)
+        , mpScope()
+        , mOutputPath(std::move(r_outputPath))
+    {
+        // "Reserve" thread map to avoid bucket moving later on.
+        const auto number_of_threads = std::thread::hardware_concurrency();
+        std::vector<std::thread> threads;
+        threads.reserve(number_of_threads);
+        std::atomic<std::size_t> thread_counter = 0;
+        for (std::size_t i_thread=0; i_thread<number_of_threads; ++i_thread) {
+            threads.emplace_back([i_thread, &thread_counter, this](){
+                while (thread_counter < i_thread) {} // <== wait until the previous thread finishes
+                mItemContainerMap.emplace(std::this_thread::get_id(), std::list<Item> {});
+                ++thread_counter;
+            });
+        }
+        for (auto& r_thread : threads)
+            r_thread.join();
+        threads.clear();
+
+        // Measure the total lifetime of the profiler
+        mpScope.reset(new Scope(mItem));
+    }
 
     Profiler(Profiler&& rOther) = default;
 
-    Profiler(std::filesystem::path&& r_outputPath);
-
-    ~Profiler();
+    ~Profiler()
+    {
+        std::ofstream file(mOutputPath);
+        file << *this;
+    }
 
     Profiler& operator=(Profiler&& rOther) = default;
 
-    [[nodiscard]] Item& Create(CodeLocation&& rItem);
+    [[nodiscard]] Item& Create(CodeLocation&& rItem)
+    {
+        auto& r_list = mItemContainerMap[std::this_thread::get_id()];
+        r_list.emplace_back(std::move(rItem));
+        return r_list.back();
+    }
 
-    [[nodiscard]] Scope Profile(Item& rItem);
+    [[nodiscard]] Scope Profile(Item& rItem)
+    {
+        return Scope(rItem);
+    }
 
     /// @brief Collect results from all threads into a single map.
-    ItemMap Aggregate() const;
+    ItemMap Aggregate() const
+    {
+        KRATOS_TRY
 
-    void Write(std::ostream& rStream) const;
+        // Aggregate maps from all threads
+        ItemMap output;
+        for (const auto& r_threadMapPair : mItemContainerMap) {
+            for (const auto& r_item : r_threadMapPair.second) {
+                auto it = output.find(r_item.mLocation);
+                if (it == output.end()) {
+                    it = output.emplace(r_item.mLocation, r_item).first;
+                } else {
+                    it->second += r_item;
+                }
+            } // for item in vector
+        } // for (location, vector) in map
+
+        return output;
+
+        KRATOS_CATCH("")
+    }
+
+    void Write(std::ostream& rStream) const
+    {
+        KRATOS_TRY
+
+        // Time the profiler's scope without changing its state
+        Item profiler_item(this->mItem);
+        --profiler_item.mCallCount;
+        --profiler_item.mRecursionLevel;
+        {Scope(profiler_item, this->mpScope->mBegin);} // Force update the copied item
+
+        auto aggregate_map = this->Aggregate();
+
+        // Sort items based on their total duration
+        std::vector<const Item*> items;
+        items.reserve(aggregate_map.size());
+        for (const auto& r_pair : aggregate_map) {
+            items.push_back(&r_pair.second);
+        }
+
+        std::sort(items.begin(),
+                  items.end(),
+                  [](const auto& rpLeft, const auto& rpRight)
+                    {return rpLeft->mCumulative < rpRight->mCumulative;});
+
+        // Start with metadata
+        Parameters root;
+        {
+            Parameters object;
+            object.AddString("name", "total");
+            object.AddString("timeUnit", GetTimeUnit<T>());
+            object.AddInt("total", profiler_item.mCumulative.count());
+            root.AddValue("meta", std::move(object));
+        }
+
+        // Add all items
+        root.AddEmptyArray("results");
+        Parameters results = root["results"];
+        for (const typename Profiler<T>::Item* p_item : items) {
+            Parameters result;
+            const auto& r_location = p_item->mLocation;
+            result.AddString("file", std::string(r_location.GetFileName()));
+            result.AddInt("line", int(r_location.GetLineNumber()));
+            result.AddString("signature", std::string(r_location.GetFunctionName()));
+            result.AddString("function", std::string(r_location.CleanFunctionName()));
+            result.AddInt("callCount", p_item->mCallCount);
+
+            std::stringstream stream;
+            stream << std::chrono::duration_cast<TimeUnit>(p_item->mCumulative).count();
+            result.AddString("total", stream.str());
+
+            stream.str("");
+            stream << std::chrono::duration_cast<TimeUnit>(p_item->mMin).count();
+            result.AddString("min", stream.str());
+
+            stream.str("");
+            stream << std::chrono::duration_cast<TimeUnit>(p_item->mMax).count();
+            result.AddString("max", stream.str());
+
+            results.Append(result);
+        }
+
+        rStream << root.PrettyPrintJsonString() << std::endl;
+
+        KRATOS_CATCH("")
+    }
 
 private:
     Profiler(const Profiler&) = delete;
@@ -190,23 +389,49 @@ private:
     std::filesystem::path mOutputPath;
 }; // class Profiler
 
-
 template <class T>
-KRATOS_API(KRATOS_CORE) std::ostream& operator<<(std::ostream& rStream, const Profiler<T>& rProfiler);
-
+std::ostream& operator<<(std::ostream& rStream, const Profiler<T>& rProfiler)
+{
+    rProfiler.Write(rStream);
+    return rStream;
+}
 
 template <class TTimeUnit>
-class KRATOS_API(KRATOS_CORE) ProfilerSingleton
+class ProfilerSingleton
 {
 public:
-    static Profiler<TTimeUnit>& Get() noexcept;
+    static Profiler<TTimeUnit>& Get() noexcept
+    {
+        std::scoped_lock<std::mutex> lock(mMutex);
+        if (!mProfiler.has_value())
+            mProfiler.emplace();
+
+        return mProfiler.value();
+    }
 
 private:
     static std::optional<Profiler<TTimeUnit>> mProfiler;
-
     static std::mutex mMutex;
 }; // class ProfilerSingleton
 
+template <class T>
+std::optional<Profiler<T>> ProfilerSingleton<T>::mProfiler;
+
+template <class T>
+std::mutex ProfilerSingleton<T>::mMutex;
+
+// Template instantiations
+template class Profiler<std::chrono::milliseconds>;
+template class ProfilerSingleton<std::chrono::milliseconds>;
+template std::ostream& operator<<(std::ostream&, const Profiler<std::chrono::milliseconds>&);
+
+template class Profiler<std::chrono::microseconds>;
+template class ProfilerSingleton<std::chrono::microseconds>;
+template std::ostream& operator<<(std::ostream&, const Profiler<std::chrono::microseconds>&);
+
+template class Profiler<std::chrono::nanoseconds>;
+template class ProfilerSingleton<std::chrono::nanoseconds>;
+template std::ostream& operator<<(std::ostream&, const Profiler<std::chrono::nanoseconds>&);
 
 } // namespace Kratos::Internals
